integrated: bootstrap, pillar sync and deploy
