+++ lr-cli/cortx_setup/commands/cluster/config/component.py
@@ -14,18 +14,46 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
+#
+# Cortx Setup API to execute complete cluster config
 
-# Cortx Setup API for configuring cortx
 
+import os
+import socket
+from pathlib import Path
 from typing import List, Optional
+
+from cortx_setup.config import (
+    ALL_MINIONS,
+    CONFIG_PATH
+)
 from cortx_setup.commands.command import Command
-from provisioner.commands import deploy
-from provisioner.salt import local_minion_id
+from cortx_setup.commands.pillar_sync import (
+    PillarSync
+)
+from cortx_setup.commands.enclosure.refresh import (
+    RefreshEnclosureId
+)
+from cortx_setup.commands.cluster.encrypt import (
+    EncryptPillar
+)
 from cortx_setup.commands.common_utils import (
     get_provisioner_states,
     get_cortx_states
 )
 
+from provisioner.commands import (
+    bootstrap_provisioner,
+    cluster_id,
+    create_service_user,
+    reset_machine_id,
+    confstore_export,
+    deploy
+)
+from provisioner.salt import (
+    local_minion_id,
+    StatesApplier
+)
 node_id = local_minion_id()
 
 noncortx_components = get_provisioner_states()
@@ -34,49 +62,176 @@ cortx_components = get_cortx_states()
 
 
 class ClusterConfigComponent(Command):
+
     _args = {
+        'nodes': {
+            'type': str,
+            'nargs': '+',
+            'optional': False,
+            'help': ('List of FQDN of node(s) to be clustered and '
+                     'bootstrapped, primary node given first.')
+        },
+        'config_path': {
+            'type': str,
+            'optional': True,
+            'default': CONFIG_PATH,
+            'help': 'Config file path for bootstrap.'
+        },
+        'source': {
+            'type': str,
+            'optional': True,
+            'default': 'rpm',
+            'help': ('Source of build to use for bootstrap. '
+                     'e.g: {local|rpm|iso}')
+        },
+        'dist_type': {
+            'type': str,
+            'optional': True,
+            'default': 'bundle',
+            'help': 'Distribution type of build'
+        },
+        'target_build': {
+            'type': str,
+            'optional': True,
+            'help': 'Target build to bootstrap'
+        },
         'type': {
             'type': str,
             'optional': True,
             'default': None,
             'dest': 'component_group',
             'choices': ['foundation', 'iopath', 'controlpath', 'ha'],
-            'help': 'Component group'
-        }
+            'help': 'Component group to deploy'
+        },
+        'iso_cortx': {
+            'type': str,
+            'optional': True,
+            'help': 'iso file path, for iso-source deployment'
+        },
+        'iso_os': {
+            'type': str,
+            'optional': True,
+            'help': ('iso os path, for iso-source deployment. '
+                     'e.g: {/opt/isos/cortx-os-1.0.0-23.iso}')
+        },
+
     }
 
     def _configure(self, components: List, stages: Optional[List] = None):
         for component in components:
-            self.logger.info(f"Applying components.{component} on nodes")
+            self.logger.debug(f"Applying components.{component} on nodes")
             try:
                 deploy.Deploy()._apply_state(
                     f"components.{component}",
                     stages=stages if stages else None
                 )
-            except Exception as ex:
-                raise ex
+            except Exception as exc:
+                raise str(exc)
 
-    def run(self, component_group: str = None):
+    def run(self, **kwargs):
         """
-        Cortx cluster config command
+        cortx cluster config command
 
-        Deploy system, prereqs and cortx components
+        Bootstrap system, deploy cortx components
 
         Execution:
-        `cortx_setup cluster config component --type <component_group>`
+        `cortx_setup cluster config component [nodes_fqdn]`
 
         """
+        try:
+            self.logger.debug(
+              "Checking for basic details in place."
+            )
+            local_fqdn = socket.gethostname()
+            nodes = kwargs['nodes']
+            config_path = Path(kwargs['config_path'])
+            target_build = kwargs['target_build']
+            component_group = kwargs['component_group']
 
-        # Deploy system, prereqs and cortx components
-        if component_group is None:
-            for component_group in noncortx_components:
-                self.logger.info(f"Deploying {component_group} components on nodes")
+            # Parsing nodes
+            for idx, node in enumerate(nodes):
+                if node == local_fqdn:
+                    nodes[idx] = f"srvnode-1:{node}"
+                else:
+                    nodes[idx] = f"srvnode-{idx+1}:{node}"
+
+            # Config validation
+            if not (CONFIG_PATH.exists() or config_path.exists()):
+                raise ValueError(f"'{config_path}' not found. "
+                                 "Please provide valid config path in command.")
+
+            # Build validation
+            if not target_build:
+                target_build = os.environ.get('CORTX_RELEASE_REPO')
+                kwargs['target_build'] = target_build
+                if not target_build:
+                    raise ValueError("'target_build' is mandatory to bootstrap. "
+                                     "Please provide valid build URL in command.")
+
+            # ISO files validation
+            if kwargs['source'] == 'iso':
+                if not (kwargs['iso_cortx'] or kwargs['iso_os']):
+                    raise ValueError(
+                         "iso single file and iso os file paths are mandatory "
+                         "to bootstrap. Please provide valid paths in command.")
+
+            self.logger.info(
+              "Initial checks done. Follow logs for progress. "
+              f"Starting bootstrap process now with args: {kwargs}"
+            )
+            bootstrap_provisioner.BootstrapProvisioner()._run(**kwargs)
+
+#            # Currently, some changes are needed for this step
+#            post_provisioner.PostProvisioner().run(**kwargs)
+
+            self.logger.debug(
+              "Starting with preparing environment. "
+              "Syncing config data now.."
+            )
+            PillarSync().run()
+
+            self.logger.debug("Generating cluster")
+            StatesApplier.apply(
+                  ["components.provisioner.config"], targets=ALL_MINIONS
+            )
+
+            self.logger.debug("Creating service user")
+            create_service_user.CreateServiceUser.run(user="cortxub")
+
+            self.logger.debug("Refreshing machine id on the system")
+            reset_machine_id.ResetMachineId.run()
+
+            self.logger.debug("Refreshing enclosure id on the system")
+            RefreshEnclosureId().run()
+
+            self.logger.debug("Setting up Cluster ID on the system")
+            cluster_id.ClusterId().run()
+
+            self.logger.debug("Encrypting config data")
+            EncryptPillar().run()
+
+            self.logger.debug("Exporting to Confstore")
+            confstore_export.ConfStoreExport().run()
+
+            self.logger.info("Bootstrap done. Proceeding to deploy components..")
+            self.logger.debug("Deploy system, prereqs and cortx components")
+
+            if component_group is None:
+                for component_group in noncortx_components:
+                    self.logger.debug(f"Deploying {component_group} components on nodes")
+                    self._configure(
+                        noncortx_components[component_group]
+                    )
+            elif component_group in cortx_components:
+                self.logger.debug(f"Deploying {component_group} components on nodes")
                 self._configure(
-                    noncortx_components[component_group]
+                    cortx_components[component_group],
+                    stages=['config.config', 'config.init_mod']
                 )
-        elif component_group in cortx_components:
-            self.logger.info(f"Deploying {component_group} components on nodes")
-            self._configure(
-                cortx_components[component_group],
-                stages=['config.config', 'config.init_mod']
+
+            return f"Bootstrapped and Deployed node(s): {nodes}"
+
+        except ValueError as exc:
+            raise ValueError(
+              f"Cluster Config Failed. Reason: {str(exc)}"
             )
